requirements.txt
#python==3.10
aiogram==3.17
asyncpg==0.29.0
bs4==0.0.2
beautifulsoup4==4.13.3
requests==2.25.1
python-decouple==3.8
setuptools==65.5.0
undetected-chromedriver==3.5.5